+++ src/Gateway/Application/TrueCode.Gateway.UseCases/Interfaces/ICurrencyRepository.cs
@@ -2,5 +2,5 @@
 
 public interface ICurrencyRepository
 {
-    Task<Dictionary<string, double>> GetFavoritesRate(long userId);
+    Task<Dictionary<string, decimal>> GetFavoritesRate(long userId);
 }
